+++ src/components/new-landing-page/index.tsx
@@ -5,7 +5,7 @@ import useHeader from './useHeader'
 import GetUpdates from '../shared/get-updates'
 import Hero from './hero'
 import Leaders from './leaders'
-import Causes from '../landing-page/causes'
+import Causes from './causes'
 // import Contributors from './contributors'
 import SocialMediaBar from '../shared/social-media-bar'
 import useExplicitHeightOnIPhone from '../../effects/useExplicitHeightOnIPhone'
